Allow more text in input dialog field.
The amount of text that can be entered in the input dialog is currently
capped at whatever can be displayed at a time, i.e. a few characters less
than the console width. This is often insufficient when a torrent should
be added via URL at standard terminal sizes.

This commit introduces paging for the input text field. The input string
is split into pages, each as long as the width of the text field. The
"page" displayed is dependant on cursor position. This only affects text
and cursor drawing, i.e. only a certain portion of the original string
will be passed to addstr.