Allow selection in peer and tracker list.
The peer and tracker panel seem to be designed to make scrolling "past"
the last element in the list impossible, i.e. the list always stops
scolling when the last list element is in view on the bottom of the
screen. This, however, makes selection of items very difficult in
scenarios where all items fit one screen, as scrollpos_detaillist is
only updated when actual screen scrolling would occur.

Remove those restrictions and let the drawing code decide how scrolled
content should be placed. In any case, scrollpos_detaillist should now
always be the index for the currently selected list item.