Scroll peer list line-by-line, prevent selection.
srollpos_detaillist will now be capped at the last index that still uses
all available display space for peer and pieces list view. This may
prevent the last items from being "selected", so this should be kept in
mind when functionality should be added in the future that acts on
selected peer.