Focus next torrent after removing another.
Resetting the focus each time a torrent is removed quickly
becomes a hassle when trying to remove multiple torrents.

This commit changes the focus to the torrent after the previously
removed one.

When removing the last torrent, self.focused_id is set to the torrent
that is about to be removed. This shouldn't be a problem, however,
as follow_list_focus() ignores invalid focused_id values.