Dialog text overflow for wide-character strings.
Dialog text sometimes overflows when it consists mainly of wide
characters (i.e. kanji). The "Remove torrent" dialog is a good example.
This commit calculates the correct text width and increases the dialog's
width as approriate. Text is also cut at the correct position if the
dialog is too small to fit the whole string.