Add method to query tracker status for a torrent.
get_tracker_responding will check each of a torrent tracker's
status and return True if at least one tracker successfully
announced recently. It will always return True if the torrent
does not have any trackers (in which case DHT will be used?).

The function checks either announce or scrape, depending on
which call was made most recently. This should normally yield the
most up-to-date status of the tracker.

If neither scrape nor announce have been performed, it returns
True, assuming that the tracker is up (until scrape/announce is
done).