Message dialog could crash when too many lines were printed.
The message window created by window() does not check for its
own height when adding lines of text. Because of this, too many lines
in conjunction with a small window can lead to program crash. The help
dialog is a good example.

Instead of writing outside of the window's allocated height, inform the
user that the window contains more text (by adding "More..." to the
dialog border on the lower left) and stop writing more text.