Open focused file with whatever is appropriate
While in the "Files" view of a torrent's Details, hitting enter on a
focused file will execute the following command, in a nonblocking
manner:

    xdg-open <focused file's path>

This will, of course, not work if the client is not launched from the
same machine as the server, because it isn't the same file structure.
If, that is the case, the operation will fail silently. On the other
hand, if the user has the current torrent files on the "client" machine
with the same path as the server, he may be amazed as how
"transmission-remote-cli" figured out where his file was. :)