Search file in other locations before opening it
Since some users prefer to save incomplete files in a separate directory
('incomplete-dir-enabled' set to true), this will make sure to check
that directory as well. In addition, it will also append '.part' to the
filename if the user has the 'rename-partial-files' enabled in
transmission.

No check is made if those options are actually enabled, because I
thought it would add unneeded complexity. I hope nothing bad will
happen.